📝 :: Rename - SoundSelectView를 SoundDetailVeiw로 변경
