+++ RelaxOn/App/Views/SoundSaveView.swift
@@ -8,13 +8,14 @@
 import SwiftUI
 
 struct SoundSaveView: View {
+    var fileInfo: FileInfo
     var body: some View {
-        Text("SoundSaveView")
+        Text("SoundSaveView \(fileInfo.name)")
     }
 }
 
 struct ThirdView_Previews: PreviewProvider {
     static var previews: some View {
-        SoundSaveView()
+        SoundSaveView(fileInfo: FileInfo(name: "세번째"))
     }
 }
